requirements.txt
Django==2.2
mysqlclient
requests==2.19.1
celery==4.4.2
redis==3.2.0
kombu==4.6.8
django-redis==4.10.0
django-celery-results==1.0.1
django-celery-beat==2.0.0
djangorestframework==3.8.2
django-filter==2.0.0
django-cors-middleware==1.3.1
drf-yasg==1.17.1
dj-database-url==0.5.0
django-constance[database]
markdown==3.0.1
sentry-sdk
kafka-python
urllib3~=1.23
paho-mqtt~=1.5.1
